public/requirements.txt
Flask==0.10.1
PyGithub
Whoosh
boto3
